Complete and test Relay Lamp status component
Why: Need to show the status of relays connected to nodes

* ...

Completed building the RelayLamp compenent by resolving errors in object
references in the data being returned by the Mission Control api.

Renamed relayNumber variables to relayNum to be more consistent with
other variables and keynames of the similar type.

changed radio button to be a checkbox. I found that browsers wanted to
handle the "only one radio button checked in a group" type of behavior

multiple checkboxes can be checked, which simplified the dispaly of
multiple 'on' relays.

* ...